requirements.txt
# MCP Server dependencies
mcp-python>=0.0.6
websockets>=10.0

# API client dependencies
requests>=2.28.0
pydantic>=2.0.0

# MongoDB dependencies
motor>=3.3.0
pymongo>=4.5.0

# Async HTTP client
aiohttp>=3.8.4

# Utilities
python-dotenv>=1.0.0 